Casting a pointer to int loses bits

+++ src/bflib_fmvids.cpp
@@ -864,7 +864,7 @@ long anim_make_FLI_BRUN(unsigned char *screenbuf) {
 		}
 	}
 	// Make the block size even
-	if ((int)animation.field_C & 1) {
+	if ((size_t)animation.field_C & 1) {
 		*animation.field_C='\0';
 		animation.field_C++;
 	}
@@ -1006,7 +1006,7 @@ long anim_make_FLI_SS2(unsigned char *curdat, unsigned char *prvdat)
 		(*lines_count)--;
 	}
 	// Make the data size even
-	animation.field_C = (unsigned char *)(((unsigned int)animation.field_C + 1) & 0xFFFFFFFE);
+	animation.field_C = (unsigned char *)(((size_t)animation.field_C + 1) & 0xFFFFFFFE);
 	return animation.field_C - blk_begin;
 }
 
@@ -1159,7 +1159,7 @@ long anim_make_FLI_LC(unsigned char *curdat, unsigned char *prvdat)
 		animation.field_C++;
 	}
 	// Make the data size even
-	animation.field_C = (unsigned char *)(((unsigned int)animation.field_C + 1) & 0xFFFFFFFE);
+	animation.field_C = (unsigned char *)(((size_t)animation.field_C + 1) & 0xFFFFFFFE);
 	return animation.field_C - blk_begin;
 }
